refactor progress
