+++ test/tfvar_input_map_test.go
@@ -1,0 +1,10 @@
+package test
+
+import (
+	"testing"
+)
+
+// TestTFVarInputMap runs a tfapply and uses the output variables as inputs to another TFApply.
+func TestTFVarInputMap(t *testing.T) {
+
+}
